rename: com.suwiki.database -> com.suwiki.security
