+++ core/security/src/androidTest/java/com/suwiki/security/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.suwiki.database
+package com.suwiki.security
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
